ビルド時に発生するNextAuthのエラーと/worldと/world/createのエラーの解消 PrismaのSchemaの不整合の解消
