+++ prisma/schema.prisma
@@ -1,9 +1,3 @@
-// This is your Prisma schema file,
-// learn more about it in the docs: https://pris.ly/d/prisma-schema
-
-// Looking for ways to speed up your queries, or scale easily with your serverless or edge functions?
-// Try Prisma Accelerate: https://pris.ly/cli/accelerate-init
-
 generator client {
   provider = "prisma-client-js"
 }
@@ -14,21 +8,20 @@ datasource db {
 }
 
 model Post {
-  id Int @id @default(autoincrement())
-  image String?
-  title String
-  content String
-  createdAt DateTime @default(now())
+  id          Int      @id @default(autoincrement())
+  title       String
+  content     String
+  createdAt   DateTime @default(now())
   countryName String?
-  userId Int
-  user User @relation(fields: [userId], references: [id])
+  userId      Int
+  user        User     @relation(fields: [userId], references: [id])
 }
 
 model User {
-  id Int @id @default(autoincrement())
-  email String @unique
+  id       Int     @id @default(autoincrement())
+  name     String?
+  image    String
+  email    String @unique
   password String
-  name String?
-  image String
-  posts Post[]
+  posts    Post[]
 }
